Started working on the ros part. Created a listening node that uses functions from the library
